Move querysets and query_expressions out of utils
They're not utilities.
